requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

openai>=1.13.3,<2.0.0

pydantic>=2.6,<3.0
pyyaml==6.0.1

redis==5.0.1

langchain>=0.1.14,<0.2.0
langchain-community>=0.0.30,<0.1.0
chromadb==0.4.24

python-dotenv==1.0.0

crewai==0.28.8
